Adds status tests
